Positioning, main and sensors
Tried to generalise the positioning library by adding an array of filters in stead of separately declared ones, added som function-calls to main in order to start WCET-analysis and made some small modifications to sensor libraries

+++ src/lib/core/io-mock.c
@@ -259,12 +259,12 @@ void analog_read_setpin(apin_t pin) {
 
 uint8_t EEP_ARRAY[EEPROM_SIZE];
 
-void eeprom_write(uint8_t p, uint8_t value)
+void eeprom_write(uint16_t p, uint8_t value)
 {
     EEP_ARRAY[p] = value;
 }
 
-uint8_t eeprom_read(uint8_t p)
+uint8_t eeprom_read(uint16_t p)
 {
     return EEP_ARRAY[p];
 }
